refactor: ImageEditor 디렉토리 이동
- 공통 컴포넌트로 사용하기 위해 상위 디렉토리로 이동
- import문을 보다 간단하게 쓸 수 있도록 index.js 작성
