+++ src/components/ImageEditor/EditorFooter.scss
@@ -1,4 +1,4 @@
-@import '../../../../stylesheets/util.scss';
+@import '../../stylesheets/util.scss';
 
 .wrapper {
   display: flex;
