+++ src/components/ImageEditor/index.js
@@ -1,0 +1,2 @@
+export { default } from './ImageEditor';
+export { UPDATE_IMAGE } from './ImageEditorLogic';
