+++ src/components/PostUploader/ImageUploader/index.js
@@ -1,0 +1,7 @@
+export { default } from './ImageUploader';
+export {
+  GET_IMAGES,
+  ADD_IMAGES,
+  DELETE_IMAGE,
+  UPDATE_REPRESENTATIVE
+} from './ImageUploaderLogic';
